update
during the respoinsive design the drop down was not working properly and with the deadline I set for myself there is just no time to continue playing with it

Now there is just a navigation and I am working on getting the first video in the background.

This video will eventually change, but for now the page will focus on sports.