update img/ html/ css
now all of the navs link to there page

it is working as far as a single scroll goes. it however is not working as far as a responsive website goes.

I have also added stuff into the photos page, the thing that needs to be learned now is how to do a cross dissolve or plave more than one circle in on this page.

I think a cross dissolve will be smartest so i can take out and plug in other photos as i continue to take more.