update css & html
fixed three of the photos and they fit the way that I want them too.

Now i have to figure out how to get the bottom three underneath the top three.

Besides that I am working on the text box now so that I can set that to the right and have it where the other text boxes are.