skeleton FINISHED
So the website is complete with no animation or fancy layout. This is what it is for right now, I am making it so that you are able to see the videos on public since I cannot seem to get the reels implemented.

The footer is completely done

I actually like the way that the video section looks.

I am hoping to get these other projects done soon do the edits and add them to the site.

Now this week will be spent focusing on making it responsive since my reset code does not seem to be working, while also starting to work on those gallery goals a little bit more.

Which will come down to editing the size of all photos probably.